use binary_switch platform
